+++ lib/src/types.dart
@@ -1,0 +1,1 @@
+typedef OnChangedString = void Function(String);