+++ CHANGELOG.md
@@ -1,7 +1,13 @@
-### Pre - [0.0.2]
+# Changelog
+
+## Pre - [0.0.3]
+
+- feat: custom text field
+
+## Pre - [0.0.2]
 
 - refactor: checkbox component
   
-### Pre - [0.0.1]
+## Pre - [0.0.1]
 
 - initial package setup